Create category flow (#10)
* Renamed transactions status to user state. Removed code related to groups.

* Added create category command. Added SetState. Added redis nil validation.

* Added create category to repository and service. Added state condition and process state.

* Updated help message

* Updated ids from int to int64.

+++ internal/domain/category.go
@@ -1,8 +1,8 @@
 package domain
 
 type CategoryDTO struct {
-	ID            int    `json:"id"`
-	UserID        int    `json:"user_id"`
+	ID            int64  `json:"id"`
+	UserID        int64  `json:"user_id"`
 	Name          string `json:"name"`
 	SanitizedName string `json:"sanitized_name"`
 	Emoji         string `json:"emoji"`
