+++ internal/service/wallets.go
@@ -11,7 +11,7 @@ type WalletsService interface {
 	Create(userID int64, name string, balance float64, timestamp *time.Time, token string) (*domain.WalletDTO, error)
 	GetAll(userID int64) (*[]domain.WalletDTO, error)
 	GetByName(name string, userID int64) (*domain.WalletDTO, error)
-	GetByID(ID int, userID int64) (*domain.WalletDTO, error)
+	GetByID(ID int64, userID int64) (*domain.WalletDTO, error)
 }
 type walletsService struct {
 	repo repository.WalletsRepository
@@ -45,7 +45,7 @@ func (s *walletsService) GetByName(name string, userID int64) (*domain.WalletDTO
 
 	return s.repo.GetByName(name, token)
 }
-func (s *walletsService) GetByID(ID int, userID int64) (*domain.WalletDTO, error) {
+func (s *walletsService) GetByID(ID int64, userID int64) (*domain.WalletDTO, error) {
 	token := jwt.GenerateToken(nil, &jwt.Payload{
 		Subject: userID,
 	})
